+++ packages/experiments/src/components/PersonaCoin/examples/PersonaCoin.Example.tsx
@@ -3,12 +3,20 @@ import { Stack, Text } from 'office-ui-fabric-react';
 import { PersonaTestImages } from '@uifabric/experiments/lib/common/TestImages';
 import { PersonaCoin } from '../index';
 
-const sectionGap = 32;
-const headingGap = 16;
-const personaCoinGap = 12;
+const tokens = {
+  sectionStack: {
+    childrenGap: 32
+  },
+  headingStack: {
+    childrenGap: 16
+  },
+  personaCoinStack: {
+    childrenGap: 12
+  }
+};
 
 const PersonaCoinStack = (props: { children: JSX.Element[] | JSX.Element }) => (
-  <Stack horizontal disableShrink gap={personaCoinGap}>
+  <Stack horizontal disableShrink tokens={tokens.personaCoinStack}>
     {props.children}
   </Stack>
 );
@@ -16,9 +24,9 @@ const PersonaCoinStack = (props: { children: JSX.Element[] | JSX.Element }) => (
 export class PersonaCoinExample extends React.Component<{}, {}> {
   public render(): JSX.Element {
     return (
-      <Stack gap={sectionGap}>
-        <Stack gap={headingGap} padding={8}>
-          <Stack gap={personaCoinGap}>
+      <Stack tokens={tokens.sectionStack}>
+        <Stack tokens={tokens.headingStack} padding={8}>
+          <Stack tokens={tokens.personaCoinStack}>
             <Text>When passing text initials will be extracted from the text</Text>
             <PersonaCoinStack>
               <PersonaCoin text="Kevin Jameson" />
@@ -28,14 +36,14 @@ export class PersonaCoinExample extends React.Component<{}, {}> {
               <PersonaCoin text="Kevin Jameson" imageUrl={PersonaTestImages.personMale} />
             </PersonaCoinStack>
           </Stack>
-          <Stack gap={personaCoinGap}>
+          <Stack tokens={tokens.personaCoinStack}>
             <Text>When passing specific initials</Text>
             <PersonaCoinStack>
               <PersonaCoin initials="JB" />
               <PersonaCoin initials="王力" />
             </PersonaCoinStack>
           </Stack>
-          <Stack gap={personaCoinGap}>
+          <Stack tokens={tokens.personaCoinStack}>
             <Text>Initials not available</Text>
             <PersonaCoinStack>
               <PersonaCoin />