+++ packages/office-ui-fabric-react/src/components/Check/Check.base.tsx
@@ -1,5 +1,4 @@
 import * as React from 'react';
-import { BaseComponent } from '../../Utilities';
 import { ICheckProps } from './Check.types';
 import { Icon } from '../../Icon';
 import { classNamesFunction } from '../../Utilities';
@@ -7,14 +6,10 @@ import { ICheckStyleProps, ICheckStyles } from './Check.types';
 
 const getClassNames = classNamesFunction<ICheckStyleProps, ICheckStyles>();
 
-export class CheckBase extends BaseComponent<ICheckProps, {}> {
+export class CheckBase extends React.PureComponent<ICheckProps, {}> {
   public static defaultProps: ICheckProps = {
     checked: false
   };
-
-  public shouldComponentUpdate(newProps: ICheckProps): boolean {
-    return this.props.checked !== newProps.checked || this.props.theme !== newProps.theme || this.props.className !== newProps.className;
-  }
 
   public render(): JSX.Element {
     const { checked, className, theme, styles } = this.props;