+++ packages/office-ui-fabric-react/src/components/FocusTrapZone/examples/FocusTrapZone.Box.Example.tsx
@@ -1,97 +1,67 @@
 import * as React from 'react';
 
-import { DefaultButton } from 'office-ui-fabric-react/lib/Button';
 import { FocusTrapZone } from 'office-ui-fabric-react/lib/FocusTrapZone';
 import { Link } from 'office-ui-fabric-react/lib/Link';
 import { TextField } from 'office-ui-fabric-react/lib/TextField';
 import { Toggle, IToggle } from 'office-ui-fabric-react/lib/Toggle';
-import './FocusTrapZone.Box.Example.scss';
+import { Stack } from 'office-ui-fabric-react/lib/Stack';
 
-export interface IBoxExampleExampleState {
-  isChecked: boolean;
+export interface IFocusTrapZoneBoxExampleState {
+  useTrapZone: boolean;
 }
 
-export default class BoxExample extends React.Component<React.HTMLAttributes<HTMLDivElement>, IBoxExampleExampleState> {
-  private _toggle: IToggle;
+export class FocusTrapZoneBoxExample extends React.Component<{}, IFocusTrapZoneBoxExampleState> {
+  public state: IFocusTrapZoneBoxExampleState = {
+    useTrapZone: false
+  };
 
-  constructor(props: React.HTMLAttributes<HTMLDivElement>) {
-    super(props);
-
-    this.state = {
-      isChecked: false
-    };
-  }
+  private _toggle = React.createRef<IToggle>();
 
   public render() {
-    const { isChecked } = this.state;
-
     return (
       <div>
-        <DefaultButton secondaryText="Focuses inside the FocusTrapZone" onClick={this._onButtonClickHandler} text="Go to Trap Zone" />
-
-        {(() => {
-          if (isChecked) {
-            return <FocusTrapZone>{this._internalContents()}</FocusTrapZone>;
-          } else {
-            return <div>{this._internalContents()}</div>;
-          }
-        })()}
+        {this.state.useTrapZone ? (
+          <FocusTrapZone>{this._internalContents()}</FocusTrapZone>
+        ) : (
+          // prettier-ignore
+          this._internalContents()
+        )}
       </div>
     );
   }
 
   private _internalContents() {
-    const { isChecked } = this.state;
+    const { useTrapZone } = this.state;
 
     return (
-      <div className="ms-FocusTrapZoneBoxExample">
-        <TextField label="Default TextField" placeholder="Input inside Focus Trap Zone" className="" />
-        <Link href="" className="">
-          Hyperlink inside FocusTrapZone
-        </Link>
-        <br />
-        <br />
+      <Stack
+        horizontalAlign="start"
+        tokens={{ childrenGap: 15 }}
+        styles={{
+          root: { border: `2px solid ${useTrapZone ? '#ababab' : 'transparent'}`, padding: 10 }
+        }}
+      >
         <Toggle
-          componentRef={this._setRef}
-          checked={isChecked}
+          label="Use trap zone"
+          componentRef={this._toggle}
+          checked={useTrapZone}
           onChange={this._onFocusTrapZoneToggleChanged}
-          label="Focus Trap Zone"
-          onText="On"
+          onText="On (toggle to exit)"
           offText="Off"
         />
-        {(() => {
-          if (isChecked) {
-            return (
-              <DefaultButton secondaryText="Exit Focus Trap Zone" onClick={this._onExitButtonClickHandler} text="Exit Focus Trap Zone" />
-            );
-          }
-        })()}
-      </div>
+        <TextField label="Input inside trap zone" styles={{ root: { width: 300 } }} />
+        <Link href="https://bing.com">Hyperlink inside trap zone</Link>
+      </Stack>
     );
   }
 
-  private _onButtonClickHandler = (): void => {
-    this.setState({
-      isChecked: true
+  private _onFocusTrapZoneToggleChanged = (ev: React.MouseEvent<HTMLElement>, checked?: boolean): void => {
+    this.setState({ useTrapZone: !!checked }, () => {
+      // Restore focus to toggle after disabling the trap zone
+      // (the trap zone itself will handle initial focus when it's enabled)
+      if (!checked) {
+        this._toggle.current!.focus();
+      }
     });
-  };
-
-  private _onExitButtonClickHandler = (): void => {
-    this.setState({
-      isChecked: false
-    });
-  };
-
-  private _onFocusTrapZoneToggleChanged = (ev: React.MouseEvent<HTMLElement>, isChecked: boolean): void => {
-    this.setState(
-      {
-        isChecked: isChecked
-      },
-      () => this._toggle.focus()
-    );
-  };
-
-  private _setRef = (toggle: IToggle): void => {
-    this._toggle = toggle;
   };
 }