+++ packages/office-ui-fabric-react/src/components/Stack/StackItem/StackItem.tsx
@@ -2,7 +2,7 @@
 import * as React from 'react';
 import { withSlots, createComponent, getSlots } from '../../../Foundation';
 import { IStackItemComponent, IStackItemProps, IStackItemSlots } from './StackItem.types';
-import { styles } from './StackItem.styles';
+import { StackItemStyles as styles } from './StackItem.styles';
 
 const view: IStackItemComponent['view'] = props => {
   const { children } = props;