+++ packages/utilities/src/customizations/Customizer.tsx
@@ -1,5 +1,4 @@
 import * as React from 'react';
-import { BaseComponent } from '../BaseComponent';
 import { Customizations } from './Customizations';
 import { CustomizerContext, ICustomizerContext } from './CustomizerContext';
 import { mergeCustomizations } from './mergeCustomizations';
@@ -19,7 +18,7 @@ import { ICustomizerProps } from './Customizer.types';
  *
  * @public
  */
-export class Customizer extends BaseComponent<ICustomizerProps> {
+export class Customizer extends React.Component<ICustomizerProps> {
   public componentDidMount(): void {
     Customizations.observe(this._onCustomizationChange);
   }