+++ apps/fabric-website-resources/CHANGELOG.md
@@ -1,6 +1,27 @@
 # Change Log - @uifabric/fabric-website-resources
 
-This log was last generated on Wed, 27 Mar 2019 12:34:02 GMT and should not be manually modified.
+This log was last generated on Tue, 02 Apr 2019 00:38:14 GMT and should not be manually modified.
+
+## 6.14.1
+Tue, 02 Apr 2019 00:38:14 GMT
+
+### Patches
+
+- Use ^ ranges instead of >=
+
+## 6.14.0
+Mon, 01 Apr 2019 12:37:03 GMT
+
+### Minor changes
+
+- add Separator page
+
+## 6.13.2
+Fri, 29 Mar 2019 12:36:45 GMT
+
+### Patches
+
+- DemoPage: pass all extra props to ComponentPage
 
 ## 6.13.1
 Wed, 27 Mar 2019 12:34:02 GMT