+++ apps/fabric-website/CHANGELOG.md
@@ -1,6 +1,39 @@
 # Change Log - @uifabric/fabric-website
 
-This log was last generated on Thu, 28 Mar 2019 12:36:39 GMT and should not be manually modified.
+This log was last generated on Thu, 04 Apr 2019 12:34:57 GMT and should not be manually modified.
+
+## 6.13.2
+Thu, 04 Apr 2019 12:34:57 GMT
+
+### Patches
+
+- HomePage: check sessionStorage for fabricVer
+
+## 6.13.1
+Tue, 02 Apr 2019 00:38:14 GMT
+
+### Patches
+
+- Use ^ ranges instead of >=
+
+## 6.13.0
+Mon, 01 Apr 2019 12:37:03 GMT
+
+### Minor changes
+
+- add Separator page
+
+### Patches
+
+- move Text under Utilities
+
+## 6.12.6
+Fri, 29 Mar 2019 12:36:45 GMT
+
+### Patches
+
+- Add session storage to index to preserve fabricVer between page navigation
+- Fix key errors from Nav
 
 ## 6.12.5
 Thu, 28 Mar 2019 12:36:39 GMT