+++ packages/experiments/CHANGELOG.md
@@ -1,6 +1,71 @@
 # Change Log - @uifabric/experiments
 
-This log was last generated on Wed, 27 Mar 2019 12:34:02 GMT and should not be manually modified.
+This log was last generated on Wed, 10 Apr 2019 12:33:17 GMT and should not be manually modified.
+
+## 6.71.0
+Wed, 10 Apr 2019 12:33:17 GMT
+
+### Minor changes
+
+- Button: Creating ActionButton variant and cleaning styling.
+
+## 6.70.0
+Tue, 09 Apr 2019 12:35:20 GMT
+
+### Minor changes
+
+- Button: Adding DefaultButton and PrimaryButton variants.
+
+## 6.69.0
+Mon, 08 Apr 2019 12:37:15 GMT
+
+### Minor changes
+
+- Button: Supporting high contrast in Button, MenuButton and SplitButton.
+
+## 6.68.0
+Fri, 05 Apr 2019 22:09:48 GMT
+
+### Minor changes
+
+- Button: Adding ariaLabel prop and updating examples to fix accessibility issues.
+
+## 6.67.1
+Fri, 05 Apr 2019 12:36:12 GMT
+
+### Patches
+
+- Use styles instead of className on theming example dialog
+
+## 6.67.0
+Wed, 03 Apr 2019 12:38:51 GMT
+
+### Minor changes
+
+- Button: Improvements on Button docs.
+
+### Patches
+
+- Button: Fixing SplitButton styling issues and correctly placing divider.
+
+## 6.66.1
+Tue, 02 Apr 2019 00:38:14 GMT
+
+### Patches
+
+- Use ^ ranges instead of >=
+- add proxy component for experiments/Separator
+
+## 6.66.0
+Mon, 01 Apr 2019 12:37:03 GMT
+
+### Minor changes
+
+- Separator: remove from experiments and promote to OUFR
+
+### Patches
+
+- Updating examples using Stack to use childrenGap token instead of gap prop.
 
 ## 6.65.0
 Wed, 27 Mar 2019 12:34:02 GMT