+++ packages/test-utilities/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log - @uifabric/test-utilities
 
-This log was last generated on Wed, 27 Mar 2019 12:34:02 GMT and should not be manually modified.
+This log was last generated on Tue, 02 Apr 2019 00:38:15 GMT and should not be manually modified.
+
+## 6.0.3
+Tue, 02 Apr 2019 00:38:15 GMT
+
+### Patches
+
+- Use ^ ranges instead of >=
 
 ## 6.0.2
 Wed, 27 Mar 2019 12:34:02 GMT